+++ src/index.d.ts
@@ -1,3 +1,0 @@
-declare module 'ketbi-tools' {
-  export function createCircleCursor(): void;
-}