+++ src/models/CursorCssInterface.ts
@@ -1,0 +1,4 @@
+export interface CursorCssInterface {
+    cursor: HTMLElement;
+    initializeCursor(): void;
+  }